Added PIDGroup to libprocess.
Log network is not specific to replicated log and the name 'network'
is also confusing. Therefore, PIDGroup is added to libprocess as a
replacement for network. This is a step towards modularizing the
PIDGroup abstraction alongside the `MasterContender` and
`MasterDetector`.

Note: The PIDGroup abstraction is a component that discovers all masters
in the group; whereas the `MasterDetector` only discovers the leader
of the group. Both are necessary for the replicated log.

Review: https://reviews.apache.org/r/50491